+++ multiversxetl/service/__init__.py
@@ -1,5 +1,5 @@
-from multiversxetl.service.mx_blocks_service import MultiversXBlocksService
+from multiversxetl.service.blocks_service import MultiversXBlocksService
 
 __all__ = [
-    "MultiversXBlocksService",
+    "MultiversXBlocksService"
 ]
